vold: Support unmount <path> detach
This is to be used primarily (exclusively?) in recovery.  In prior
versions, the local file was copied to tmpfs so that everything could
be unmounted cleanly.  But the current version uses the local file
in-place due to constraints with low memory devices.  This prevents
the filesystem that holds the local file from being unmounted in the
normal way.  So in order to avoid leaking a mount, the filesystem that
holds the local file must be unmounted with the detach flag.

Change-Id: I17716f131225b54ca3bc9d2215a45d22d1841137